pyproject.toml
[project]
name = "openrelik-worker-ip-extractor"
version = "0.1.0"
description = "OpenRelik worker that extracts IPv4/IPv6 addresses from Windows EVTX event logs."
authors = [{ name = "Julian Hill", email = "julian.hill@ncsc.nl" }]
requires-python = ">=3.11,<3.12"

# Dependencies installed into the worker image
dependencies = [
  "celery>=5.3",
  "python-evtx>=0.7.4",
  "defusedxml>=0.7.1",
  "lxml>=5.2.2",
  "orjson>=3.10.7",
  "openrelik-worker-common>=0.1.0",  # shared utils (logging, file_utils, etc.)
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "pytest-cov>=5.0",
  "mypy>=1.11",
  "ruff>=0.6.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
addopts = "-q"

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.11"
strict = false

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.build.targets.sdist]
include = ["src", "pyproject.toml", "openrelik.yaml", "README.md"]